Split into multiple scripts
Not working version though
